requirements.txt
flask
google-api-python-client>=2.100.0
google-auth>=2.28.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
openai>=1.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
requests>=2.31.0
twilio>=9.0.0
watchdog>=3.0.0
python-dateutil>=2.8.2